Add transactions; Fix bindings in Profile and Profile Edit;

+++ app/src/main/java/domain/model/TravelPreview.kt
@@ -1,12 +1,12 @@
 package domain.model
 
 data class TravelPreview(
-    val id: Int,
+    val id: Long,
     val name: String,
     val image: String,
 ) {
 companion object {
-    const val ADD_ID = 0
+    const val ADD_ID = 0L
     fun getAddItem() = TravelPreview(ADD_ID, "", "")
     fun List<TravelPreview>.cloned() = ArrayList(this.map { it.copy() })
 }
